pyproject.toml
[project]
name = "eduzenbot"
description = "A bot for telegram"
readme = "README.md"
version = "v0.beta"
requires-python = ">=3.13"
keywords = []
authors = [
  { name = "Eduardo Enriquez (eduzen)", email = "eduardo.a.enriquez@gmail.com" },
]
dependencies = [
  "attrs",
  "beautifulsoup4",
  "cachetools",
  "emoji",
  "peewee",
  "python-telegram-bot",
  "requests",
  "sentry-sdk",
  "tmdbsimple",
  "tweepy",
  "yfinance",
  "python-dotenv",
  "logfire>=2.11.1",
]

[dependency-groups]
dev = [
  "pre-commit>=4.0.1",
  "ipython",
  "ipdb",
  "factory-boy",
  "pytest",
  "coverage",
  "pytest-mock",
  "pytest-xdist",
  "pytest-recording",
  "pytest-randomly",
  "mypy",
  "types-beautifulsoup4",
  "types-cachetools",
  "types-colorama",
  "types-docutils",
  "types-Pygments",
  "types-psycopg2",
  "types-pytz",
  "types-requests",
  "types-setuptools"
]

[tool.isort]
profile = "black"

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "eduzenbot/__about__.py",
  "eduzenbot/scripts/*",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.coverage.json]
pretty_print=true
